requirements.txt
black==23.3.0
click==8.1.3
colorama==0.4.6
keyboard==0.13.5
MouseInfo==0.1.3
mss==8.0.2
mypy-extensions==1.0.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
PyAutoGUI==0.9.53
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
PyYAML==6.0
